requirements.txt
asn1crypto==0.22.0
astroid==1.5.2
cffi==1.10.0
cryptography==1.9
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.1
django-bootstrap3==8.2.3
djangorestframework==3.6.3
Fabric3==1.13.1.post1
gunicorn==19.7.1
idna==2.5
isort==4.2.5
lazy-object-proxy==1.3.1
mccabe==0.6.1
paramiko==2.1.2
pep8==1.7.0
pyasn1==0.2.3
pycparser==2.17
pylint==1.7.1
pytz==2017.2
selenium==3.4.3
six==1.10.0
static3==0.7.0
wrapt==1.10.10
